# Use IReadOnlyList type-matching to avoid enumerator allocation.